fix (infrastructure): replace shell based bootstrap implementation (WIP)
kafka topics creation should happen in the  managed environnement of a kafka container because the scripts actually needs the artifacts created during kafka build to run, at a given position in the filesystem. Such a constraint forces to do a lot of work for this to work in a portable way using shell scripts. The better solution would be to spin up an other kafka container that's behaving as the client to request le kafka cluster